+++ packages/plugin-packages/mashroom-portal-demo-solidjs-app/src/js/modules.d.ts
@@ -1,0 +1,5 @@
+// Declare modules without existing TypeScript definitions
+
+declare module '*.scss';
+
+declare module '*.svg';