+++ packages/plugin-packages/mashroom-portal-demo-solidjs-app/webpack.config.js
@@ -1,0 +1,68 @@
+const ESLintPlugin = require('eslint-webpack-plugin');
+
+module.exports = {
+    entry: __dirname + '/src/js',
+    output: {
+        path: __dirname + '/dist',
+        filename: 'bundle.js',
+    },
+    module: {
+        rules: [
+            {
+                test: /\.(ts|tsx)$/,
+                exclude: /node_modules/,
+                use: [
+                    {
+                        loader: 'babel-loader',
+                    },
+                ],
+            },
+            {
+                test: /\.scss$/,
+                use: [
+                    {
+                        loader: 'style-loader',
+                    },
+                    {
+                        loader: 'css-loader',
+                    },
+                    {
+                        loader: 'sass-loader',
+                    },
+                ],
+                sideEffects: true,
+            },
+            {
+                test: /\.svg$/,
+                use: [
+                    {
+                        loader: 'svg-inline-loader',
+                    },
+                ],
+            }
+        ],
+    },
+    externals: [],
+    resolve: {
+        extensions: ['.js', '.ts', '.tsx'],
+    },
+    plugins: [
+        new ESLintPlugin({
+            extensions: ['.js', '.ts', '.tsx'],
+            fix: true,
+        })
+    ],
+    devServer: {
+        host: '0.0.0.0',
+        allowedHosts: 'all',
+        port: 8094,
+        static: 'src',
+        open: true,
+        client: {
+            overlay: {
+                errors: true,
+                warnings: false,
+            },
+        }
+    },
+};