+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ## [unreleased]
 
+ * Added a [SolidJS](https://www.solidjs.com) demo Portal App (Microfrontend)
  * Dropped support for IE11 and ES5
  * Upgrade to CodeMirror 6 with autocomplete support in the CSS editor
  * Admin Toolbar: Cleanup the DOM properly after a drag ends